Add `helm-init` script.
`helm-init` clones a repo and creates a branch locally & remotely so we can
begin work on a ticket.

It mimics the git clone command, but we only the master branch and the branch we
need. And by using git init, we set up hooks automatically if `init.templatedir`
is set in the .gitconfig