helm: add *all* remote branches; only fetch master.
Allow all the remote branches to be fetched/tracked, not just `master`.

The '-t' option is not specified when adding the 'origin' remote.
Previously, when the repo was setup, the master and ONLY the master was added to
the remotes:

    git remote add -f -t master -m master origin $URL
    # produces refspec: fetch = +refs/heads/master:refs/remotes/origin/master

This means that the branch created/fetched later in the script for the ticket
could not be fetched and tracked because git was told to track master and only
master.

Setting the `-t` option to `master` meant 'add this remote and only track the
branch named master'. `-t` can take multiple branches as well (`-t master dev
feature`). By removing `-t`, we use git's default wildcard refspec:

    fetch = +refs/heads/*:refs/remotes/origin/*

See:
http://git-scm.com/docs/git-remote.html
http://git-scm.com/docs/git-branch.html
http://git-scm.com/docs/git-fetch
http://git-scm.com/book/en/Git-Internals-The-Refspec

Or manual entries:
man git-remote
man git-branch
man git-fetch